May 3 , 2013
*Essentially Tutorial 4 Complete - With errors - *

**
At the moment the game can currently run with all arrows keys working,
so you can run the game as an application and test the game yourself.

I've run into a problem that I currently have no idea how to fix.
You'll notice when you jump on the first platform and jump again the
game freezes up and you'll receive an error.

In his tutorial he does not encounter this issue, so obviously
a mistake has been made. I spent quite a bit of time trying to figure it
out, but I was unsuccessful.

I would advise you to go through Part 2,3 and 4 of the tutorials and
follow along with the code that i've already typed up and see if you can
find an error.

I Noticed that earlier in the videos he set tween = 0.07,
and then later on I totally missed that he added a function called
setTween and changed tween = 1, so a little confusing.