May 5,2013(2)
**Part 6 Of Game Tutorial Complete** --With Minor error.

So part 6 is complete. This is the most I'm going to get done before
next week. I probably won't be able to touch the game until Monday night
or Tuesday at some point.

a brief update of whats been done up to this point is as followed:
--Player can now explore the map
--Player can glide
--Player can throw fireballs
--Player can scratch
--Enemy slug (Test enemy) has been added and can hit walls
and animation will update to turn around and to load all the
different sprites (Walk cycle)
--Player HUD has been added with Health and FirePower in place

**Error occurs when enemy fires a fireball, firepower should
empty, and then replentish, but as it appears the firePower
doesn't empty, rather it increases beyond the boundry (26/25, 27/25
etc..), so obviously something is going backwards and I'm not to sure at
this point where it may be.

Once again, you should catch up and watch the tutorials and follow along
with the code to fix any errors I may have missed (made) throughout. And
if you do fix the problem with the firepower,
please feel free to continue to Part 7.

We should take full advantage of the fact that these videos aren't
coming out everyday and try and brainstorm what unique gameplay we'd
like to make for our own original game.

Please don't fall off the planet on me again, lol.